Resources 追加、3D Object の凝視実装
