+++ Assets/MyProject/Material/MusicUI.mat.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
-DefaultImporter:
+guid: fe509bfe8748d4cb499be06349c65991
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 0
   userData: 
   assetBundleName: 
   assetBundleVariant:
