+++ Assets/MyProject/Prefabs/Background.prefab.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
-DefaultImporter:
+guid: 115419b3d46d3478cad4d884042d0160
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
