+++ Assets/MyProject/Prefabs/VirtualCameraMock.prefab
@@ -10,7 +10,6 @@ GameObject:
   m_Component:
   - component: {fileID: 1466598320843669723}
   - component: {fileID: 1924460603510967183}
-  - component: {fileID: 5127196062308098304}
   m_Layer: 0
   m_Name: MainCamera_R
   m_TagString: EditorOnly
@@ -75,14 +74,6 @@ Camera:
   m_OcclusionCulling: 1
   m_StereoConvergence: 10
   m_StereoSeparation: 0.022
---- !u!81 &5127196062308098304
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 3663077782246275224}
-  m_Enabled: 1
 --- !u!1 &6389071016953284916
 GameObject:
   m_ObjectHideFlags: 0
@@ -142,7 +133,6 @@ GameObject:
   m_Component:
   - component: {fileID: 6389071017859788214}
   - component: {fileID: 6389071017859788208}
-  - component: {fileID: 6389071017859788215}
   m_Layer: 0
   m_Name: MainCamera_L
   m_TagString: EditorOnly
@@ -207,14 +197,6 @@ Camera:
   m_OcclusionCulling: 1
   m_StereoConvergence: 10
   m_StereoSeparation: 0.022
---- !u!81 &6389071017859788215
-AudioListener:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 6389071017859788213}
-  m_Enabled: 1
 --- !u!1 &7294614521208569696
 GameObject:
   m_ObjectHideFlags: 0
