+++ Assets/MyProject/Scenes/00_Init.unity
@@ -830,7 +830,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 0
+  m_IsActive: 1
 --- !u!114 &2058040984
 MonoBehaviour:
   m_ObjectHideFlags: 0
