+++ Assets/MyProject/Scenes/21_Game.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: bba1e625d54e34001ac52371fbb2e608
 DefaultImporter:
   externalObjects: {}
   userData:
