+++ Assets/MyProject/Scripts/Domain/Common/IButtonImpl.cs.meta
