+++ Assets/MyProject/Scripts/Domain/Virtual/ValueObject.cs
@@ -3,7 +3,8 @@
     public struct Value
     {
         public const string VirtualStateKeyValue = "VirtualState";
-        public const int ObjectUILayer = 1 << 10;
+        public const int CanvasUILayer = 1 << 5;
+        public const int ObjectUILayer = 1 << 11;
     }
 
     public enum VirtualState
