+++ Assets/MyProject/Scripts/UI/Scene/SceneRoot.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a2c7de3649bb14e4b97a2bc97a2f4de4
+guid: 4b82bcad382374f00a3ed27098867d76
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
