+++ Assets/MyProject/Scripts/UI/Virtual/VirtualGazeObjectControl.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a2c7de3649bb14e4b97a2bc97a2f4de4
+guid: 2a888ac8679344f43ab715e93a12a203
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
