+++ Assets/Resources/Prefabs/Manager.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 5102097357261250745}
   - component: {fileID: 5102097357261250750}
   - component: {fileID: 7240365868369945364}
+  - component: {fileID: 343230429753661590}
   m_Layer: 0
   m_Name: Manager
   m_TagString: Untagged
@@ -56,6 +57,18 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 89cfaf614d845439ba55c0f0552f2749, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+--- !u!114 &343230429753661590
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 5102097357261250747}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a888ac8679344f43ab715e93a12a203, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
 --- !u!1001 &5102097358651791968
