+++ Assets/VRSampleScenes.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: d519c63ccdb5bed4f977067267a3b55e
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
