+++ Assets/VRSampleScenes/Materials.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: 008f76b63d3ad0543a3e42137771e65a
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
