+++ Assets/VRSampleScenes/Materials/Menu.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: df1086e961acc944dbed589985b52299
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
