+++ Assets/VRSampleScenes/Materials/Menu/MenuBg.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f65c96498c277e429931be96cf3682c
+timeCreated: 1439479010
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
