+++ Assets/VRSampleScenes/Materials/Utils.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: 4c288cc0e8db1e8479d4b1d185db6588
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
