+++ Assets/VRSampleScenes/Materials/Utils/BlobShadow.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 41f3fb97620cda2479bf1b170b998c0c
+timeCreated: 1447332146
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
