+++ Assets/VRSampleScenes/Models.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: 0bd001f994412034a8f17068ad0b7f69
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
