+++ Assets/VRSampleScenes/Models/Menu.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: c79e1a05043753246856c5e8ab50211c
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
