+++ Assets/VRSampleScenes/Prefabs.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: 438a5e729d170f04eb3c334065b6b021
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
