+++ Assets/VRSampleScenes/Prefabs/Utils.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: 427ed006b6e14764c99086452acf4dff
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
