+++ Assets/VRSampleScenes/Prefabs/Utils/BlobShadow.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 2dda78a7affe63a4d81e05200cc3429d
+timeCreated: 1447332346
+licenseType: Store
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
