+++ Assets/VRSampleScenes/Textures.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: bb72f10720df7bb4691fd47d64a3efe2
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
