+++ Assets/VRSampleScenes/Textures/Shared.meta
@@ -1,5 +1,6 @@
 fileFormatVersion: 2
-guid: d91fcabeed50c4e719a36f56bbf42004
+guid: 52a419e3e7f777f408d6d1fe4954def5
+folderAsset: yes
 DefaultImporter:
   externalObjects: {}
   userData:
